+++ tests/cypress/e2e/datasets.cy.js
@@ -1,5 +1,4 @@
-import { retryableBefore } from "../support/retryableBefore.js"
-import { stringToArray, randomInteger } from "../support/utils.js"
+import { retryableBefore, stringToArray, randomInteger } from "../support/utils.js"
 
 /**
  * List of dataset ids
