+++ tests/cypress/e2e/portallogin.cy.js
@@ -1,4 +1,4 @@
-import { retryableBefore } from "../support/retryableBefore.js"
+import { retryableBefore } from "../support/utils.js"
 
 describe('Portal Login', { testIsolation: false }, function () {
     retryableBefore(function () {
